Refactor and clean up
Introduce a Manipulator class that is responsible for higher level changes to the Bitmap.
This seperates responsibilities and leaves the Bitmap responsible for holding state and low level changes - just setting individual pixels.
In future the Manipulator class could be further broken down if it gets more complex.

Also introduce a Response object the the so all Manipulator actions have a command response.
This way the calling code does not need to worry about the implementation, it just expects an object that responds to message to be returned.